courses/advanced-programming/workshops/workshop_4/requirements.txt
uvicorn
sqlalchemy
fastapi
psycopg2-binary
python-dotenv